requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
alembic==1.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
